add --output option; refactor tests
