+++ tests/atest/transformers/DiscardEmptySections/test.py
@@ -1,0 +1,24 @@
+from .. import run_tidy_and_compare
+
+
+class TestDiscardEmptySections:
+    TRANSFORMER_NAME = 'DiscardEmptySections'
+
+    def test_removes_empty_sections(self):
+        run_tidy_and_compare(self.TRANSFORMER_NAME, source='removes_empty_sections.robot')
+
+    def test_removes_empty_sections_except_comments(self):
+        run_tidy_and_compare(
+            self.TRANSFORMER_NAME,
+            source='removes_empty_sections.robot',
+            expected='removes_empty_sections_except_comments.robot',
+            config=':allow_only_comments=True'
+        )
+
+    def test_remove_selected_empty_node(self):
+        run_tidy_and_compare(
+            self.TRANSFORMER_NAME,
+            source='removes_empty_sections.robot',
+            expected='removes_selected_empty_section.robot',
+            config=' --startline 17 --endline 18'
+        )
